+++ esp-hal/build.rs
@@ -89,10 +89,15 @@ fn main() -> Result<(), Box<dyn Error>> {
                 "(ESP32 only) Enables a workaround for the issue where SPI in half-duplex mode incorrectly transmits the address on a single line if the data buffer is empty.",
             ),
             (
-                "place-more-rodata-in-ram",
+                "place-switch-tables-in-ram",
                 Value::Bool(false),
-                "Places more data (lookup-tables, switch-tables etc.) into RAM - resulting in better performance (especially for interrupts) but more RAM consumption",
-            )
+                "Places switch-tables, some lookup tables and constants related to interrupt handling into RAM - resulting in better performance but slightly more RAM consumption.",
+            ),
+            (
+                "place-anon-in-ram",
+                Value::Bool(false),
+                "Places anonymous symbols into RAM - resulting in better performance at the cost of significant more RAM consumption. Best to be combined with `place-switch-tables-in-ram`.",
+            ),
         ],
         true,
     );